+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
 
+## 7.9
+- PyMast routing
+* PyMast Reroute gui
+- query.has_link_to
+- follow route to force a route e.g. force science scan
+- Routes now check so they don't schedule multiple times (For selections)
+
+
 ## 7.8
 
 - Made Mast globals illegal as variable name e.g. max = 2 is an error (Added test)